requirements.txt
Werkzeug==0.11.2
Flask==0.10.1
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
flask-discoverer==0.0.2
Flask-Migrate==1.6.0
requests==2.8.1
httpretty==0.8.12
SQLAlchemy==1.0.10
psycopg2==2.6.1
ConcurrentLogHandler==0.9.1
consulate==0.6.0
flask-consulate==0.1.2
testing.postgresql==1.3.0
nosecomplete==0.1.1
nltk==3.2.1
textblob==0.11.1
